requirements.txt
alembic==1.4.2
aniso8601==8.0.0
autopep8==1.5.2
bcrypt==3.1.7
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-HTTPAuth==4.0.0
Flask-JWT==0.3.2
Flask-JWT-Extended==3.24.1
flask-marshmallow==0.13.0
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.3
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.6.1
marshmallow-sqlalchemy==0.23.1
oauthlib==2.1.0
passlib==1.7.2
psycopg2-binary==2.8.5
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
SQLAlchemy==1.3.17
toml==0.10.1
urllib3==1.25.9
Werkzeug==1.0.1
uwsgi
